misc/HtmlTestSuite/results/test-0605.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>CSS Fonts Module Level 3: parsing font-variant-numeric with invalid values</title>
<link href="https://www.w3.org/TR/css-fonts-3/#font-variant-numeric-prop" rel="help"/>
<meta content="font-variant-numeric supports only the grammar 'normal | [ &lt;numeric-figure-values&gt; || &lt;numeric-spacing-values&gt; || &lt;numeric-fraction-values&gt; || ordinal || slashed-zero ]'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
</head>
<body>
<script>
test_invalid_value('font-variant-numeric', 'auto');

test_invalid_value('font-variant-numeric', 'normal lining-nums');
test_invalid_value('font-variant-numeric', 'lining-nums oldstyle-nums');

test_invalid_value('font-variant-numeric', 'proportional-nums normal');
test_invalid_value('font-variant-numeric', 'tabular-nums proportional-nums');

test_invalid_value('font-variant-numeric', 'normal diagonal-fractions');
test_invalid_value('font-variant-numeric', 'diagonal-fractions stacked-fractions');

test_invalid_value('font-variant-numeric', 'ordinal normal');

test_invalid_value('font-variant-numeric', 'normal slashed-zero');
</script>
</body>
</html>
